requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
azure-core==1.32.0
azure-identity==1.20.0
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
dependency-injector==4.44.0
Deprecated==1.2.18
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fastapi_msal==2.1.6
fastjsonschema==2.21.1
flake8==7.2.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
microsoft-kiota-abstractions==1.9.2
microsoft-kiota-authentication-azure==1.9.2
microsoft-kiota-http==1.9.2
microsoft-kiota-serialization-form==1.9.2
microsoft-kiota-serialization-json==1.9.2
microsoft-kiota-serialization-multipart==1.9.2
microsoft-kiota-serialization-text==1.9.2
mistune==3.1.0
msal==1.32.0
msal-extensions==1.2.0
msgraph-core==1.3.2
msgraph-sdk==1.28.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
ngrok==1.4.0
numpy==2.2.1
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.13
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.3.0
pure_eval==0.2.3
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.10.3
pydantic-extra-types==2.10.1
pydantic-settings==2.8.1
pydantic_core==2.27.1
pyflakes==3.3.2
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
scipy==1.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
std-uritemplate==2.0.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
win32_setctime==1.2.0
wrapt==1.17.2
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
